Add Ingest::EPG::Date and plan pending tests

+++ db/migrate/20180220163350_remove_services_id_from_schedule.rb
@@ -1,5 +1,5 @@
 class RemoveServicesIdFromSchedule < ActiveRecord::Migration[5.1]
   def change
-    remove_column :schedules, :services_id
+    remove_column :schedules, :services_id if Schedule.column_names.include?("services_id")
   end
 end
